feat: Edit/Detail basic User
