+++ pages/dashboard/users/[id].tsx
@@ -1,5 +1,10 @@
 import { useRouter } from 'next/router'
 import React from "react"
+import UserForm from "../../../components/User/UserForm";
+import {
+    Button,
+} from '@material-ui/core';
+import ConfirmDialog from '../../../components/Dialogs/ConfirmDialog';
 
 const users = [
     { id: 1, email: "email1@mail.com", lastName: 'Snow', name: 'Jon', role: "manager" },
@@ -18,16 +23,50 @@ const UserDetail = () => {
     const { id } = router.query
     const [currentUser, setCurrentUser] = React.useState(null)
 
+    const [mode, setMode] = React.useState("view");
+    const [openDeleteModal, setOpenDeleteModal] = React.useState(false);
+    
     React.useEffect(() => {
         let user = users.find(user =>  user.id === parseInt(id as string))
         if(user) setCurrentUser(user)
         else setCurrentUser(null)
     }, [])
 
+    const onDelete = () => {
+        console.log("Delete user and redirect to users");
+        router.push("/dashboard/users")
+    }
+
     return (
         <>
             <h1>User Detail: {id}</h1>
-            <p>{currentUser && currentUser.email}</p>
+            
+            {mode === 'view' && (
+                <>
+                    <Button variant="contained" color="primary" onClick={() => setMode('edit')}>
+                        Edit
+                    </Button>
+                    <Button variant="contained" color="secondary" onClick={() => setOpenDeleteModal(true)}>
+                        Delete
+                    </Button>
+                </>    
+            )}
+            {mode === 'edit' && (
+                <Button variant="contained" color="primary" onClick={() => setMode('view')}>
+                    Cancel
+                </Button>    
+            )}
+            <UserForm mode={mode} />
+            {/* TODO: show user bookings */}
+
+            <ConfirmDialog
+                title="Delete User?"
+                open={openDeleteModal}
+                setOpen={setOpenDeleteModal}
+                onConfirm={onDelete}
+            >
+                Are you sure you want to delete this user?
+            </ConfirmDialog>
         </>    
     )
 }
